+++ packages/react-charts/src/components/ChartTooltip/index.ts
@@ -1,3 +1,1 @@
-export * from './ChartCursorFlyout';
-export * from './ChartCursorTooltip';
 export * from './ChartTooltip';
